android [nfc]: Add ReactExtensions.kt
From MainActivity.kt and from notifications/NotifyReact.kt.

We'll move some more of this kind of React-extending code here soon.
